requirements.txt
numpy==1.22.4
Pillow==10.4.0
torch==1.10.1+cu111
torchvision==0.11.2